# swagger/index.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API"
  },
  "host": "localhost:10054",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "accessToken": {
      "type": "apiKey",
      "name": "authorization",
      "scheme": "bearer",
      "in": "header"
    },
    "refreshToken": {
      "type": "apiKey",
      "name": "refresh",
      "in": "header"
    }
  },
  "paths": {
...
    "/notice/{id}": {
      "get": {
        "tags": [
          "공지사항 관련(게시판)"
        ],
        "summary": "특정공지 호출",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "responses": {},
        "security": [
          {
            "accessToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "공지사항 관련(게시판)"
        ],
        "summary": "특정공지 수정",
        "consumes": "multipart/form-data",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          },
          {
            "in": "formData",
            "name": "notice",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "title",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "content",
            "type": "TEXT"
          }
        ],
        "responses": {},
        "security": [
          {
            "accessToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "공지사항 관련(게시판)"
        ],
        "summary": "공지 삭제",
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "responses": {},
        "security": [
          {
            "accessToken": []
          }
        ]
      }
    }
  }
}